Moved gpu octree tests to test/gpu. (#3965)
* Moved gpu octree tests to test/gpu.

* add .points and std:: after removal of previous PRs.

* exclude gpu octree tests from ctest

* add cmake option for excluding gpu tests
