+++ test/gpu/CMakeLists.txt
@@ -1,0 +1,5 @@
+if(NOT HAVE_CUDA)
+  return()
+endif()
+
+add_subdirectory(octree)
