+++ test/gpu/octree/data_source.hpp
@@ -131,7 +131,7 @@ struct DataGenerator
             float query_radius = radius > 0 ? radius : radiuses[i];
             const PointType& query = queries[i];
 
-            for(std::size_t ind = 0; ind < size(); ++ind)
+            for(std::size_t ind = 0; ind < points.size(); ++ind)
             {
                 const PointType& point = points[ind];
